S-63270 Minimal passing Acceptance Test!
We took a large cut through an even simpler test that referenced in the story. We queried a single Story by its OID, and then printed its OID token. That's it.

In doing this, we discovered that we were able to completely reuse the existing V1Connector.GetData function without needing to introduce RestSharp at all. This is a testament to the work Acey and Velocity Partners did a couple years ago. Well done. All we had to do was ask for application/json, and then parse it with NewtonSoft to dramatically simplify accessing the data. No need to parse through XML now.